Support .abc files directly for Arnold standin look assignment
